requirements.txt
fastapi
uvicorn
google-genai          
google-cloud-firestore 
google-api-python-client 
google-auth-oauthlib   
google-auth           
python-multipart      
pydantic